requirements.txt
chainlit==0.4.101
cohere==4.11.2
hydra-core==1.3.2
InstructorEmbedding==1.0.1
langchain==0.0.214
numpy==1.25.0
openai==0.27.8
pandas==2.0.0
pinecone-client==2.2.2
python-dotenv==1.0.0
sentence-transformers==2.2.2